requirements.txt
requests
lxml
PyYAML